+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+### 1.2.8
+1.StickyTabs tabs子元素兼容，使用4.23.0前antd，TabPane形式
+### 1.2.7
+1.README.md fix
+### 1.2.6
+1.AutoSizeList解决item宽度不一致问题，改为使用grid布局
+### 1.2.5
+1.AutoSizeList加入ref getRowCount, 去除ref rowCount
+### 1.2.4
+1.AutoSizeList useImpretiveHandle deps
+### 1.2.3
+1.stickyTabs优化，AutoSizeList加入ref rowCount
+### 1.2.2
+1.stickyTabs 加入offsetTop属性
+### 1.2.1
+1.tabs items属性需要antd >= 4.23.0，更改peerDependencies
+### 1.2.0
+1.form增加rangePicker类型
 ### 1.1.10
 1.更改git地址
 2.更改邮箱